// src/databasepatcher-web/src/main/frontend/app/dashboard/models/Database.ts
export class Database {
    name: string;
    hostname: string;
    port: string;
    username: string;
    password: string;
    lastPatchTimestamp: string;
    lastStatus: string;
    currentVersion: string;
    nextPatchTimestamp: string;
    targetVersion: string;
    region: string;
    zone: string;
    type: string;
    regionId: string;
    availabilityZoneId: string;
    dataTypeId: string;
    id: string;
    patchId: string;
    runNow: string;


}